pyproject.toml
[tool.poetry]
name = "aesop"
version = "0.1.0"
description = "A generative AI Youtube Empire"
authors = ["TH3-F001 <123153982+TH3-F001@users.noreply.github.com>"]
license = "GPL3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.22.0"
tldextract = "^5.1.2"
google-auth = "^2.32.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.136.0"
oauth2client = "^4.1.3"
openai = "^1.35.10"
beautifulsoup4 = "^4.12.3"
opencv-python-headless = "^4.10.0.84"
numpy = "^2.0.0"
mutagen = "^1.47.0"
ffmpeg-python = "^0.2.0"
pydub = "^0.25.1"
elevenlabs = "^1.3.1"
praw = "^7.7.1"
newsapi-python = "^0.2.7"
pygobject = "^3.48.2"
bump2version = "^1.0.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
